Adjusting the RepositoryBrowserMediator to use a recursive search function when populating the expansion of a directory as a user browses.
Adjusting the converter to use old logic to remove the accidental creation of files and directories for the same path in SVN. This is a bug and will still cause PNG or Binary files to appears as directories, but it is the lesser of two evils. 

Adding HDividedBox and VDividedBox to give a bit of flexibility to the frontend.