Creating the ability to override configurations via file system properties.
If a deployer does not specify a file in /tmp/svnanalytics.properties then the values in svnanalyticsDefault.properties will be used for JDBC Connectivity.

If a deployer creates a file in /tmp/ called svnanalytics.properties and creates key=value pairs that match those in the defaults with different values then this file will override the defaults.

If a deployer creates a file anywhere on the filesystem and uses a -DconfigOverrideFilePath=/path/to/svnanalytics.properties then this file will take priority and override all previously configured settings.