Working towards abstracting out the necessary code to allow Oracle and MySQL database support via a property that can be override at runtime.
Right now including the datasource by name is required. So long term that would mean seperate war files that are database specific could be created. 

The ideal scenario here is going to be a way to make the test cases run for both oracle and mysql as seperate test runs so that a full regression will occur to ensure database support grows properly.