SVNANALYTICS-1 SVNANALYTICS-4
Adding support to derive the root of a repository in case it is not the same as the actual repository's root
Displaying a fault string on repository save errors.
Adding logic to ensure the startRevision of a repository is set when saving a new repository at a lower level than the root for users who want to monitor / index a sub-directory of a larger repository.