COMMONSUSER-3
Adding call to ensure that when a new user is added it will appear in the grid. This will break when fully secured though since it will require a logged in user to make this call.
Adding initial definition of spring authentication-provider to hook into the message broker. This will require some refactoring because it should use a DAO Authentication provider.
Refactoring the dependency on HibernateDAOSupport since it will be removed in 3.0.0.RC1 of Spring.
Implementing Externalizable on DTOs since it is needed to properly serialize with new Flex Mojos / GAS3 DTO generation.
Defining initial security declarations on the UserService.java to dictate the level of security required to authenticate.
Defining two login/logout methods.