COMMONSUSER-3
Ensuring the login functionality actually works by adding a call to the login remote object.
Adding a check to see if the user is logged in, if not a login dialog displays. 
Also adding the ability for a user to logout.
Integrating the authenticationDao with a custom SQL for privileges and username selection. This could still use some polish as this is sort of cobbled together in this commit.
Getting rid of IExternalizable on the UserDTO as it seems to serve no purpose. 
All UserService calls are secured with ROLE_ADMIN at this time, more granular control can be added later.